+++ CHANGELOG.md
@@ -17,6 +17,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- ### Fixed -->
 <!-- ### Removed -->
 
+## Unreleased
+
+### Changed
+
+- The playground service worker now serves its own 404s when a file is not
+  found, instead of forwarding the request to the server.
+
 ## [0.3.3] - 2020-12-01
 
 ### Fixed